refactor(logs): remove direct console_manager usage from logs_cmd.py and update tests; logs_cmd.py now returns only Error/Success, no direct printing; all user-facing output handled by CLI/command_handler; updates test_cli_logs.py to match new pattern and patch correct locations; all logs subcommand CLI tests pass
